refactor: move AnimatorParser to AnimatorParsers folder
