+++ Editor/AnimatorParsers/AnimationParser.cs
@@ -7,7 +7,7 @@ using UnityEditor.Animations;
 using UnityEngine;
 using static Anatawa12.AvatarOptimizer.ErrorReporting.BuildReport;
 
-namespace Anatawa12.AvatarOptimizer.Processors.TraceAndOptimizes
+namespace Anatawa12.AvatarOptimizer.AnimatorParsers
 {
     class AnimationParser
     {
