+++ Editor/AnimatorParsers/ModificationsContainers.cs
@@ -6,7 +6,7 @@ using JetBrains.Annotations;
 using UnityEngine;
 using Object = UnityEngine.Object;
 
-namespace Anatawa12.AvatarOptimizer.Processors.TraceAndOptimizes
+namespace Anatawa12.AvatarOptimizer.AnimatorParsers
 {
     interface IModificationsContainer
     {
