+++ Editor/Processors/TraceAndOptimize/AutoFreezeBlendShape.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using Anatawa12.AvatarOptimizer.AnimatorParsers;
 using nadena.dev.ndmf;
 using UnityEditor;
 using UnityEngine;
