+++ Editor/Processors/TraceAndOptimize/GCComponentInfo.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using Anatawa12.AvatarOptimizer.AnimatorParsers;
 using JetBrains.Annotations;
 using UnityEngine;
